requirements.txt
algoliasearch-django>=4.0,<5.0
django>=5.2.1
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
